Update GCD.java
I made a mistake in my first explanation, saying that if a was 0, then the GCD between a and b would be |a|, which is zero, in this case, which is wrong (don't divide by zero). I've altered the explanation to denote that if a = 0, then the GCD of a and b should be |b|, or absolute value of b.